refactor(tracking): name the service for what it tracks
`services::tracking` did not say tracking of what, the same objection
that turned `user_integrations` into `user_media_trackers` on #240. The
module builds what a media tracker needs, so it takes that name.

The addon capability keeps `tracking`: slots there are named for the
activity, alongside `meta`, `catalog` and `metrics`.
