+++ crates/remux-server/src/services/mod.rs
@@ -1,8 +1,8 @@
 pub mod image;
+pub mod media_tracker;
 pub(crate) mod resolve;
 pub(crate) mod stream_service;
 pub mod stremio;
-pub mod tracking;
 
 pub use resolve::MediaResolveService;
 pub(crate) use resolve::ResolvedItem;
